Re-work DOI workflow
